# contracts/openapi.json
{
  "components": {
    "schemas": {
      "CredentialsRequest": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "CredentialsRequest",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FocusId": {
        "enum": [
          "technical_details",
          "research_progress",
          "major_changes",
          "breaking_events",
          "niche_trends",
          "industry_changes",
          "controversy_changes",
          "deep_context"
        ],
        "title": "FocusId",
        "type": "string"
      },
      "FocusOption": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/FocusId"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "FocusOption",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "api": {
            "const": "ok",
            "default": "ok",
            "title": "Api",
            "type": "string"
          },
          "database": {
            "const": "ok",
            "default": "ok",
            "title": "Database",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          },
          "worker": {
            "const": "ok",
            "default": "ok",
            "title": "Worker",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "InvestmentMarketId": {
        "enum": [
          "china_market",
          "us_stock",
          "crypto_market"
        ],
        "title": "InvestmentMarketId",
        "type": "string"
      },
      "InvestmentMarketOption": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/InvestmentMarketId"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "InvestmentMarketOption",
        "type": "object"
      },
      "OnboardingAnswers": {
        "properties": {
          "focus_ids": {
            "items": {
              "$ref": "#/components/schemas/FocusId"
            },
            "title": "Focus Ids",
            "type": "array"
          },
          "investment_market_ids": {
            "items": {
              "$ref": "#/components/schemas/InvestmentMarketId"
            },
            "title": "Investment Market Ids",
            "type": "array"
          },
          "scope_ids": {
            "items": {
              "$ref": "#/components/schemas/ScopeId"
            },
            "title": "Scope Ids",
            "type": "array"
          }
        },
        "required": [
          "scope_ids",
          "investment_market_ids",
          "focus_ids"
        ],
        "title": "OnboardingAnswers",
        "type": "object"
      },
      "OnboardingResponse": {
        "properties": {
          "answers": {
            "$ref": "#/components/schemas/OnboardingAnswers"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "focus_options": {
            "items": {
              "$ref": "#/components/schemas/FocusOption"
            },
            "title": "Focus Options",
            "type": "array"
          },
          "investment_market_options": {
            "items": {
              "$ref": "#/components/schemas/InvestmentMarketOption"
            },
            "title": "Investment Market Options",
            "type": "array"
          },
          "scope_options": {
            "items": {
              "$ref": "#/components/schemas/ScopeOption"
            },
            "title": "Scope Options",
            "type": "array"
          }
        },
        "required": [
          "completed",
          "scope_options",
          "investment_market_options",
          "focus_options",
          "answers"
        ],
        "title": "OnboardingResponse",
        "type": "object"
      },
      "OnboardingSelection": {
        "properties": {
          "focus_ids": {
            "items": {
              "$ref": "#/components/schemas/FocusId"
            },
            "minItems": 1,
            "title": "Focus Ids",
            "type": "array"
          },
          "investment_market_ids": {
            "items": {
              "$ref": "#/components/schemas/InvestmentMarketId"
            },
            "title": "Investment Market Ids",
            "type": "array"
          },
          "scope_ids": {
            "items": {
              "$ref": "#/components/schemas/ScopeId"
            },
            "minItems": 1,
            "title": "Scope Ids",
            "type": "array"
          }
        },
        "required": [
          "scope_ids",
          "investment_market_ids",
          "focus_ids"
        ],
        "title": "OnboardingSelection",
        "type": "object"
      },
      "ScopeId": {
        "enum": [
          "ai",
          "open_source",
          "technology",
          "science",
          "investment"
        ],
        "title": "ScopeId",
        "type": "string"
      },
      "ScopeOption": {
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ScopeId"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "ScopeOption",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/SessionState"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionUser"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "state",
          "user"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "SessionState": {
        "enum": [
          "anonymous",
          "onboarding_required",
          "ready"
        ],
        "title": "SessionState",
        "type": "string"
      },
      "SessionUser": {
        "properties": {
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "SessionUser",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Infoscope API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly is_session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly is_session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "get_health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/onboarding": {
      "get": {
        "operationId": "get_onboarding_api_v1_onboarding_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Onboarding",
        "tags": [
          "onboarding"
        ]
      },
      "put": {
        "operationId": "update_onboarding_api_v1_onboarding_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/scope": {
      "get": {
        "operationId": "get_scope_api_v1_scope_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scope",
        "tags": [
          "onboarding"
        ]
      },
      "put": {
        "operationId": "update_scope_api_v1_scope_put",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update Scope",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/v1/session": {
      "get": {
        "operationId": "get_current_session_api_v1_session_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "is_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Session",
        "tags": [
          "auth"
        ]
      }
    }
  }
}